Port chains were being returned that could not be cast to int(). Casting to float() seems to work.

+++ temperusb/temper.py
@@ -54,16 +54,16 @@ def find_ports(device):
         if matches:
             bus_str = readattr(dirent, 'busnum')
             if bus_str:
-                busnum = int(bus_str)
+                busnum = float(bus_str)
             else:
                 busnum = None
             dev_str = readattr(dirent, 'devnum')
             if dev_str:
-                devnum = int(dev_str)
+                devnum = float(dev_str)
             else:
                 devnum = None
             if busnum == bus_id and devnum == dev_id:
-                return int(matches.groups()[1])
+                return float(matches.groups()[1])
 
 
 class TemperDevice(object):
